% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dat.R
\docType{data}
\name{dat}
\alias{dat}
\title{Example data for illustrating NSEC and maNSEC calculations using a binomial response.}
\format{
A data frame with 21 rows and 3 variables:
\describe{
\item{concentration}{effluent concentration in unspecified units}
\item{response}{number of successful macroalgal germinations}
\item{n}{number of macroalga tested}
}
}
\source{
{Table 2 in Fisher and Fox (2023)}
}
\usage{
dat
}
\description{
Gives Number of successful germinations (response) out of n seeds of a species of macroalga when exposed to various concentrations of a treated effluent.
}
\examples{
data(dat)
}
\keyword{datasets}
